genesis: allow building arbitrary Zephyr applications
The build command is generally convenient for making Zephyr
applications, mainly since it allows easy out of tree builds. This can
be useful e.g. when building the Bluetooth HCI SPI application for
96b_carbon_nrf51.

To allow this, add a --skip-signature option to the build command
arguments, which bypasses the zep2newt step during application
build. Also allow any board to be specified when it's passed.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>