genesis: refactor common argument handling
To prepare for a 'genesis flash' command, pull the common arguments
out of the 'make' helpers. The flash command will need some of these
common arguments, but it won't directly invoke make, so it's cleaner
to keep this out of the make helpers.

While we're here, add whitelist support, which 'genesis flash' will
need.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>