genesis: refactor, removing "make_" helpers
Let's move the "make" helper routines into the core class we have to
share command policy in, and rename it Command in the process.

This is prep work for adding a documentation builder; the next prep
step is extending Command so it's a proper abstract base class, and
reworking the handler routines into subclasses.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>