genesis: support operations on multiple boards
Allow building binaries for, configuring, and flashing multiple boards
with the same invocation of genesis. This is accomplished by accepting
multiple -b options and building a list.

For example, to build a signed application binary for app APP_NAME, as
well as mcuboot images, for 96b_carbon, 96b_nitrogen, and frdm_k64f:

  ./genesis build -b 96b_carbon -b 96b_nitrogen -b frdm_k64f APP_NAME

To use xconfig to configure just the applications:

  ./genesis configure -C xconfig -o app \
            -b 96b_carbon -b 96b_nitrogen -b frdm_k64f APP_NAME

To flash an application to multiple boards:

  ./genesis flash -b 96b_carbon -b 96b_nitrogen APP_NAME

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>