genesis: flash: add device_id argument
The only reliable way to flash one single device when multiple devices
of the same type are available is by specifying the board unique
device id. The device id identifier is manipulated by the respective
binary flash class and converted into the right argument type and
name, which translates to serial for dfu-utils and board id for
pyocd-flashtool.

Since there is no simple way to map a device id to a board type, only
one board argument can be used when a device id is specified. The
device id argument can also be called multiple times for flashing
several devices in a row.

Signed-off-by: Ricardo Salveti <ricardo.salveti@linaro.org>