genesis.py: use prebuilt GCC ARM embedded toolchain
Rather than using a prebuilt Zephyr toolchain as planned, add support
for using a prebuilt GCC ARM embedded toolchain instead, using a
prebuilt repository which was added to the manifest.

This requires dtc to be available on the host system. That's handled
in a separate genesis-dev metapackage update.

The Zephyr SDK can still be used with this patch, but some additional
arguments are required:

./genesis build --prebuilt-toolchain=n --zephyr-gcc-variant=zephyr ...

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>