Somewhat hackish support for cmake.
Switch to using the new upstream Zephyr CMake build system, with some
hacks for flash support depending on (currently) out of tree patches.

To keep things simple and easy on OS X (and in our CI environment),
generate the host tools and add them to PATH when building binaries,
so users don't need to do it themselves (OS X) or have a Zephyr SDK
installed (Linux).

Update the 'build' command in a backwards-compatible way, using
'outdir' instead of 'build', and copying binary artifacts (.bin and
signed .bin) to the locations where CI expects to find them for now.

We need some hacks to pass values through to the flasher script (which
functionality is being proposed for upstream) via the environment.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>