commands.py: don't pad the images
Due to a misunderstanding of how MCUboot works, we've been
unnecessarily padding the images for a long time.

You can still build bootable and updatable images without padding. The
images just have to take care to mark themselves OK (in bank 0) at
boot time after passing any self-tests, and to write the magic bytes
in bank 1 when requesting an upgrade.

This change will affect applications which don't behave properly with
respect to MCUboot, but they were buggy to begin with. Furthermore,
not padding the images makes both flashing and updates go faster (only
the bytes really required by the application must be transmitted), and
has the benefit that FOTA updates are more energy and bandwidth
efficient as well.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>