commands.py: disable menuconfig build on Windows
This is currently unsupported. This commit can be reverted when Zephyr
issue 5847 is resolved (though a follow-up commit to delete
prepare_host_tools() entirely would probably be needed as well, since
the plan would be to replace a host tools build with Python tooling
that doesn't need to be compiled).

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>