Add prep-release-notes.py.
This is a reincarnated version of the get-mergeup-highlights.py script
that was formerly part of the documentation. It takes us one step
closer to automating our release notes handling by generating YAML
fragments in the versioned release notes schema format. These still
have to be copied into the right sections of the release notes file,
which must be updated based on the template.

It depends on the nonstandard 'click' module. This won't be added to
our official ZMP dependencies, since it's an internal API for OSF.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>