zephyr_mergeup.py: print out summary of OSF commits
At the end of the commit message, print out a commented block with
information on the remaining outstanding OSF commits. "Outstanding"
here means that it's a commit reachable from osf-ref but not reachable
from upstream-ref which is neither a merge nor a commit which has been
reverted. That is, it's the current set of patches which are authored
by OSF and not merged upstream.

Also use this list of outstanding commits to search for patches which
were in fact merged upstream and need to be reverted before doing the
mergeup. Do this by filtering the upstream commits for those which
have an OSF email, and seeing if any outstanding OSF patches have
shortlogs whose edit distance is small compared to one or more of
those upstream commits. This is just a heuristic, but it should catch
most cases.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>