commands.py: move HELP to top-level
The way help works with these commands was designed to try to make
things consistent between commands, but it's confusing and
clumsy.

It'd be better to move it to top level and let each command handle
itself, trusting the implementors to keep things consistent. This
trivial refactor starts the process.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>