prep-release-notes.py: use pinned manifests as source data
The current behavior of starting from a baseline commit and going
until each repository's HEAD is not correct. Update builds are
constructed from zmp-manifest builds which receive E2E treatment, and
may be older than the current HEADs.

To make this work more conveniently, ask for pinned manifests for the
previous (start) and current (end) update, and prep release notes
based on that.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>