zmp: re-work MCUboot support
zmp's current behavior with regards to building MCUboot is not exactly
what we want:

- Upstream has added a CONFIG_BOOTLOADER_MCUBOOT, which automatically
  takes care of setting up CONFIG_TEXT_SECTION_OFFSET and is the
  recommended method for requesting MCUboot support, but we don't use
  it. This prevents us from building arbitrary Zephyr sample code as
  an MCUboot child image.

- While zmp has a --skip-signature option to 'zmp build' that implies
  only building the application, its name is confusing to users, and
  there's no corresponding support for 'zmp flash' to flash an
  application built without MCUboot support.

- The --skip-signature option should be incompatible with all the
  --imgtool-xxx options, but it's currently not treated as such.

The right thing to do is use CONFIG_BOOTLOADER_MCUBOOT by default, but
have --skip-signature disable that behavior. To make it more clear
what's going on, rename --skip-signature to '--no-bootloader' (but
keep the old option around for compatibility).

Provide an overlay config file, mcuboot-overlay.conf, which sets
CONFIG_BOOTLOADER_MCUBOOT, and use it appropriately in zmp depending
on the --no-bootloader option. Pick this up in 'zmp flash' to do the
right thing. This means users can now do the following:

    # Build and flash an arbitrary Zephyr sample with mcuboot support:
    ./zmp build zephyr/samples/hello_world
    ./zmp flash zephyr/samples/hello_world

    # A clean build is currently necessary when switching between
    # mcuboot support and not:
    rm -r outdir/zephyr/samples/hello_world

    # Build and flash an arbitrary Zephyr sample without mcuboot:
    ./zmp build --no-bootloader zephyr/samples/hello_world
    ./zmp flash zephyr/samples/hello_world

While we're here, clean up the core.py app/mcuboot location code so
it's easier to read, and enforce the mutual exclusion of
--no-bootloader and the --imgtool-xxx options.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>