zephyr_whatsnew.py: tweak area summary
The markdown text will want to generate this as a list, as the table
output doesn't look very good, so prepare for keeping consistency with
that by doing it in the text formatter.

This also means we no longer have to deal with a bug in the output
padding calculations.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>