zephyr_whatsnew: factor out analysis code from output formatting
The output formatting is about to get more sophisticated to support
multiple use cases. It will be useful to have the code which is
responsible for analyzing repository history to be separated from it,
to keep the final result clean.

Add a new ZephyrRepoAnalyzer class that does all the heavy lifting of
checking out the repository history, and returns a ZephyrRepoAnalysis
containing the results when it's done. Move the analysis code into
ZephyrRepoAnalysis.

Update the output formatting code to use analysis objects instead.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>