zephyr_whatsnew.py: add per-analyzer area hook
It's often the case that a string of commits will all have the same
shortlog prefix, but that prefix isn't worth adding to
AREA_TO_SHORTLOG_RES.

Example:

    intel_s1000: Add intel_s1000 SoC
    intel_s1000: define memctl_default.S and memerror-vector.S files
    intel_s1000: create xt-sim_intel_s1000_defconfig
    intel_s1000: add intel_s1000_crb board
    intel_s1000: uart: configure UART for intel_s1000
    intel_s1000: gpio: enable GPIO handling
    intel_s1000: i2c: enable I2C for intel_s1000
    intel_s1000: tests: introduce tests to check features enabled
    intel_s1000: scripts: debug, debugserver and flash scripts

It would be nice to be able to specify a map from prefixes (like
intel_s1000) to a given area (like Boards) during a single run of the
script, but not modify the script itself.

As step one, add a ZephyrRepoAnalyzer parameter that maps shortlog
prefixes to areas on an instance-by-instance basis.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>