prep-release-notes.py: improve heuristics
Add a state machine for printing the YAML which uses heuristics to
detect:

- old and new names for the list of upstream patches

- internal sections within the highlights

- the two major styles of sections, which either include headings,
  like this:

      Some heading:

      Summary paragraph 1.

      Multiple paragraphs are allowed.

      Another heading:

      This one just has a single paragraph.

  or which don't, like this:

      First summary paragraph about this component.

      Another summary paragraph. The YAML headings are unknown and
      should be emitted as TODO.

The state machine now prints the internal sections as YAML comments,
so they don't have to be deleted manually, and automatically populates
release notes highlights with headings when they are detected.

This reduces the time spent manually processing the script output.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>