commands.py: go west, young commands
Use the newly-merged "west" meta-tool to pull build configuration out
of the runner package, and to invoke flashing with extra arguments (in
our case, those necessary to pass signed binaries) to support MCUboot
child image chainloading.

As of this commit and Zephyr v1.12.0-rc1, it is finally possible to
use zmp with a vanilla Zephyr tree, making it easier to bisect bugs.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>