commands: bring back --device-id as --board-id
Use the fact that west supports extra options at the command line to
bring back the ability to specify a board.

Rename the option --board-id, since that's the option supported by
upstream west.

We can't bring back --extra even with west, though, as the more
generalized flashing support provided by west doesn't support the idea
of a single command which is used to flash the board. For example,
some flash runners start a GDB server, connect to it with another GDB
process, and use that to flash the board.

Rather than try to support an option which isn't general enough, just
remove --extra.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>