commands.py: look for 'mcuboot.overlay' instead
The current MCUboot build will add boards/BOARD-mcuboot.overlay to the
MCUboot build as well as the application build if it exists.

That's always seemed too magical and non-obvious to me for reproducing
an application build by hand.

Let's change this bit of magic to looking for a single, top-level
'mcuboot.overlay' instead. That can always contain board-specific
logic if it needs to, and has the benefit of being more obvious to
'ls' run in the top level application directory.

Signed-off-by: Marti Bolivar <marti@foundries.io>